Moved the implementation out of the public task type
